+++ apps/kitchen-sink/components/ui/Input/creator/InputContext.tsx
@@ -1,5 +1,0 @@
-import type { InputContext } from './types';
-import { createContext } from '../../../../utils/gluestack-utils/common';
-
-export const [InputProvider, useInput] =
-  createContext<InputContext>('InputContext');