+++ apps/kitchen-sink/components/ui/accordion/aria/types.ts
@@ -1,4 +1,0 @@
-export type State = {
-  selectedValues: string[];
-  toggleItem: (itemValue: string, isDisabled?: boolean) => void;
-};