+++ apps/kitchen-sink/components/ui/radio/creator/RadioProvider.tsx
@@ -1,3 +1,0 @@
-import { createContext } from '@/utils/gluestack-utils/common';
-
-export const [RadioProvider, useRadio] = createContext<any>('RadioContext');