+++ apps/todo-app/components/ui/flat-list/index.tsx
@@ -1,2 +1,0 @@
-'use client';
-export { FlatList } from 'react-native';