+++ apps/todo-app/components/ui/radio/creator/RadioGroupContext.tsx
@@ -1,4 +1,0 @@
-import { createContext } from '@/utils/gluestack-utils/common';
-
-export const [RadioGroupProvider, useRadioGroup] =
-  createContext<any>('RadioGroupContext');