+++ apps/todo-app/components/ui/refresh-control/index.tsx
@@ -1,2 +1,0 @@
-'use client';
-export { RefreshControl } from 'react-native';