+++ apps/website/components/ui/hstack/creator/index.ts
@@ -1,9 +1,0 @@
-export function createHStack<StyledHStackProps>({
-  Root,
-}: {
-  Root: React.ComponentType<StyledHStackProps>;
-}) {
-  Root.displayName = 'HStack';
-
-  return Root;
-}