+++ apps/website/components/ui/progress/creator/ProgressContext.tsx
@@ -1,4 +1,0 @@
-import { createContext } from '@/utils/gluestack-utils/common';
-
-export const [ProgressProvider, useProgress] =
-  createContext<any>('ProgressContext');