+++ packages/utils/gluestack-utils/aria/overlays/src/useOverlayTrigger.web.ts
@@ -1,1 +1,0 @@
-export { useOverlayTrigger } from '@react-aria/overlays';