+++ scripts/mappers/website/templateGenerator.ts
@@ -18,8 +18,8 @@ export const generateCodePreviewer = (
 ) => {
   const sourcePath =
     component === 'use-break-point-value' || component === 'use-media-query'
-      ? path.resolve('packages/components/ui/utils')
-      : path.resolve('packages/components/ui');
+      ? path.resolve('src/components/ui/utils')
+      : path.resolve('src/components/ui');
   const examplePath = path.join(sourcePath, component, 'examples', exampleName);
   const codePath = path.join(examplePath, 'template.handlebars');
   const argsPath = path.join(examplePath, 'meta.json');