+++ packages/create-gluestack/CHANGELOG.md
@@ -1,5 +1,23 @@
 # create-gluecode
 
+## 1.0.11
+
+### Patch Changes
+
+- perf: optimize template cloning with shallow git clone
+
+## 1.0.9
+
+### Patch Changes
+
+- fix: fixed cleanup process after clone
+
+## 1.0.7
+
+### Patch Changes
+
+- chore: restructered codebase
+
 ## 1.0.6
 
 ### Patch Changes